+++ Userland/native/tests/test_util.c
@@ -68,7 +68,7 @@ void endless_loop() {
 }
 
 void endless_loop_print(uint64_t wait) {
-      int64_t pid = sys_get_pid();
+      int pid = sys_get_pid();
 
     while (1) {
         printf("%d ", pid);
